http: Explicitly deprecate request::content
Modern httpd handlers must use request body stream to get the contents
of the request, but this "must" is in fact "should" as there's no such
visible requirement other than a code comment nobody reads.

Http client callers are not supposed to access this member either,
there's write_body() method for that. But this time it's simpler -- the
client is still in the experimental namespace.

Still we need to handle httpd users and deprecate content explicitly.

Just marking the field as deprecated is ugly as compiler generates
deprecation warnings every time it default- or move-constructs the
request even though the content may not used explicitly. This can be
fixed by defining default and move constructors and shutting up compiler
with with necessary pragmas. Worse -- deprecation warnings pop up when
_destructing_ request even with the help of explicit destructor.

To fix the destruction challenge the `sstring content` is converted into
a reference and the "real" content becomes `private _content`. Since
references are not destructed, compiler is blissfully ignorant.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>