build: use find_dependency() instead find_package() in config file
SeastarDependencies.cmake is used in two scenarios:

- configure Seastar library, when building Seastar itself.
- configure Seastar library, when a parent project tries to
  detect Seastar library. when building the Seastar application.

in the second case, we should respect the "REQUIRED" and "QUIET"
options passed to `find_package(Seastar)`. and CMake provides
`find_dependency()` macro for this purpose.

in this change, we conditionally call `find_dependency()` when
`SeastarDependencies.cmake` is used when this script is called
when finding Seastar library. this provides a better developer
experience to Seastar developers.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>